02_02_wine_pairing/requirements.txt
langchain
langchain-openai
python-dotenv
pandas
ipython
